Add auth in controller commission and group

+++ src/main/kotlin/unq/pds/api/controller/GroupController.kt
@@ -1,12 +1,18 @@
 package unq.pds.api.controller
 
-import io.swagger.v3.oas.annotations.*
-import io.swagger.v3.oas.annotations.media.*
-import io.swagger.v3.oas.annotations.responses.*
-import org.springframework.http.*
+import io.jsonwebtoken.Jwts
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.ArraySchema
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.ExampleObject
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import unq.pds.api.dtos.MessageDTO
 import unq.pds.api.dtos.GroupDTO
+import unq.pds.api.dtos.MessageDTO
 import unq.pds.model.Group
 import unq.pds.services.GroupService
 import javax.validation.Valid
@@ -26,7 +32,7 @@ class GroupController(private val groupService: GroupService) {
         value = [
             ApiResponse(
                 responseCode = "200",
-                description = "success",
+                description = "Success",
                 content = [
                     Content(
                         mediaType = "application/json",
@@ -46,8 +52,16 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun createGroup(@RequestBody @Valid group: GroupDTO): ResponseEntity<Group> {
-        return ResponseEntity(groupService.save(group.fromDTOToModel()), HttpStatus.OK)
+    fun createGroup(@CookieValue("jwt") jwt: String?, @RequestBody @Valid group: GroupDTO): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.save(group.fromDTOToModel()), HttpStatus.OK)
     }
 
     @GetMapping
@@ -90,8 +104,16 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun getGroup(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
-        return ResponseEntity(groupService.read(id), HttpStatus.OK)
+    fun getGroup(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.read(id), HttpStatus.OK)
     }
 
     @PutMapping
@@ -134,8 +156,16 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun updateGroup(@RequestBody group: Group): ResponseEntity<Any> {
-        return ResponseEntity(groupService.update(group), HttpStatus.OK)
+    fun updateGroup(@CookieValue("jwt") jwt: String?, @RequestBody group: Group): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.update(group), HttpStatus.OK)
     }
 
     @DeleteMapping
@@ -180,7 +210,15 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun deleteGroup(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+    fun deleteGroup(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        if (body["role"] == "STUDENT") return ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
         groupService.delete(id)
         return ResponseEntity(MessageDTO("Group has been deleted successfully"), HttpStatus.OK)
     }
@@ -225,8 +263,20 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun addMember(@NotBlank @PathVariable groupId: Long, @NotBlank @PathVariable studentId: Long): ResponseEntity<Any> {
-        return ResponseEntity(groupService.addMember(groupId, studentId), HttpStatus.OK)
+    fun addMember(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable groupId: Long,
+        @NotBlank @PathVariable studentId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.addMember(groupId, studentId), HttpStatus.OK)
     }
 
     @PutMapping("/removeMember/{groupId}/{studentId}")
@@ -269,8 +319,20 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun removeMember(@NotBlank @PathVariable groupId: Long, @NotBlank @PathVariable studentId: Long): ResponseEntity<Any> {
-        return ResponseEntity(groupService.removeMember(groupId, studentId), HttpStatus.OK)
+    fun removeMember(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable groupId: Long,
+        @NotBlank @PathVariable studentId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.removeMember(groupId, studentId), HttpStatus.OK)
     }
 
     @PutMapping("/addProject/{groupId}/{projectId}")
@@ -313,8 +375,20 @@ class GroupController(private val groupService: GroupService) {
                 )]
             )]
     )
-    fun addProject(@NotBlank @PathVariable groupId: Long, @NotBlank @PathVariable projectId: Long): ResponseEntity<Any> {
-        return ResponseEntity(groupService.addProject(groupId, projectId), HttpStatus.OK)
+    fun addProject(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable groupId: Long,
+        @NotBlank @PathVariable projectId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.addProject(groupId, projectId), HttpStatus.OK)
     }
 
     @GetMapping("/getAll")
@@ -335,7 +409,15 @@ class GroupController(private val groupService: GroupService) {
                 ]
             )]
     )
-    fun getAll(): ResponseEntity<List<Group>> {
-        return ResponseEntity(groupService.readAll(), HttpStatus.OK)
+    fun getAll(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(groupService.readAll(), HttpStatus.OK)
     }
 }
